grep/pcre2: move back to thread-only PCREv2 structures
Change the setup of the "pcre2_general_context" to happen per-thread
in compile_pcre2_pattern() instead of in grep_init(), as happens with
all the rest of the pcre2_* members of the grep_pat structure.

This is done to simplify the setup, my longer term plan is to move
these regex functions to a common library outside of the grep.c code,
to e.g. use it in diffcore-pickaxe.c. Needing to deal with both global
and per-thread setup/teardown in such an API is annoying if we don't
have to.

It should also save save us memory allocations in the common case, if
micro-optimizing this case was worth it (it's not).

since now we'll only do the memory allocation on-demand. We only need
this data when we're

, and if we're trying to micro-optimize memory use
this'll save us memory for the common-case

at trivial cost of a little bit more memory use
under threading.

On my x86_64 system it's an extra 24 bytes per thread. So e.g. on an 8
core machine we'll allocate 384 bytes instead of the fixed 48.

94da9193a6e (grep: add support for PCRE v2, 2017-06-01)

In 513f2b0bbd4 (grep: make PCRE2 aware of custom allocator,
2019-10-16) we introduced the need to allocate and manage global
variables for PCREv2, in addition to the the thread-local allocations
we've done since 94da9193a6 (grep: add support for PCRE v2,
2017-06-01). The same was done for PCREv1 in 57d4660468e (grep: make
PCRE1 aware of custom allocator, 2019-10-16).

Johannes's comments:
https://public-inbox.org/git/nycvar.QRO.7.76.6.1908052120302.46@tvgsbejvaqbjf.bet/

    > instead of using a static variable inside this helper function it
    > might be better to use one extra field inside the (struct grep_pat
    > *p), where all other variables are kept

    My thinking about that was that this would add more code, and thus more
    opportunities to introduce bugs. Also, it's not like the general context
    really has any _state_ per se. It just registers the current allocator,
    which we want to assume is constant over the life-time of the process.

    So does it really make sense to create one general context per grep
    pattern? (Or even per grep pattern and thread?)